split e2e
